requirements.txt
oauthlib
pandas==0.19.2
py-flags
requests_oauthlib